From xen-unstable:

 changeset:   15970:04cc0e22a20a0bdf6ceefb016185175ee094ed30
 user:        Keir Fraser <keir@xensource.com>
 date:        Thu Sep 27 16:36:23 2007 +0100
 files:       tools/ioemu/audio/audio.c tools/ioemu/audio/mixeng.c tools/ioemu/audio/ossaudio.c tools/ioemu/block-raw.c tools/ioemu/block-vvfat.c tools/ioemu/bswap.h tools/ioemu/cutils.c tools/ioemu/monitor.c tools/ioemu/osdep.h tools/ioemu/target-i386-dm/exec-dm.c tools/ioemu/usb-linux.c tools/ioemu/vl.c tools/ioemu/vl.h tools/ioemu/vnc.c
 description:
 ioemu: Fixes for BSD.

Mostly merged.  Dropped the ctype.h casts since we have those already,
and also a few things that looked wrong.

+++ cutils.c
@@ -23,7 +23,7 @@
  */
 #include "qemu-common.h"
 
-void pstrcpy(char *buf, int buf_size, const char *str)
+void pstrcpy(char *buf, size_t buf_size, const char *str)
 {
     int c;
     char *q = buf;
@@ -41,7 +41,7 @@ void pstrcpy(char *buf, int buf_size, const char *str)
 }
 
 /* strcat and truncate. */
-char *pstrcat(char *buf, int buf_size, const char *s)
+char *pstrcat(char *buf, size_t buf_size, const char *s)
 {
     int len;
     len = strlen(buf);
